# api/openapi.yaml
openapi: 3.1.0
info:
  title: RemitX Anchor API
  version: "0.1.0"
  description: |
    RemitX Stellar anchor API implementing SEP-1, SEP-10, SEP-12, and SEP-31
    for trustless cross-border remittances on Stellar.

    ## Authentication

    Most endpoints require a **SEP-10 JWT** obtained from `POST /auth`:

    ```
    Authorization: Bearer <jwt>
    ```

    The `/api/v1/escrow/:id/release` and `/api/v1/escrow/:id/dispute` endpoints
    additionally require the JWT to be signed with an **Ed25519 key** (full
    SEP-10 Ed25519 verification, not the shared-secret variant).

    Content / tier management endpoints authenticate via the
    `X-Stellar-Account` header (Stellar G… public key).

    The admin reconciliation endpoint requires a separate admin JWT with
    `role: "admin"`.

    ## Idempotency

    All mutating endpoints (POST / PUT / PATCH / DELETE) **must** include an
    `Idempotency-Key` header (UUID v4) so that retries are safe. The server
    replays the original response for duplicate keys.

  contact:
    name: RemitX Engineering
    email: dev@remitx.io
    url: https://discord.gg/remitx
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.remitx.io
    description: Production
  - url: http://localhost:8000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
...
components:
  securitySchemes:
    sep10JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT obtained from `POST /auth` after completing the SEP-10
        challenge/response flow. Valid for 24 hours.

    adminJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Admin JWT with `role: "admin"` claim. Issued out-of-band to
        operations staff.

    stellarAccount:
      type: apiKey
      in: header
      name: X-Stellar-Account
      description: >
        Stellar Ed25519 public key (G… address) of the authenticated creator
        or fan. Used by content / tier management endpoints.

  # -------------------------------------------------------------------------
  # Reusable headers
  # -------------------------------------------------------------------------
...
  parameters:
    EscrowId:
      name: id
      in: path
      required: true
      description: UUID of the escrow record.
      schema:
        type: string
        format: uuid
        example: "a3bb189e-8bf9-3888-9912-ace4e6543002"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: "escrow not found"
        detail:
          type: string
          description: Optional extended detail (admin / server errors only).
          example: "connection refused: postgres:5432"

    # ── SEP-10 ──────────────────────────────────────────────────────────────
...
    EscrowState:
      type: string
      enum: [Funded, Released, Refundable, Refunded, Cancelled]
      description: Current state of the escrow in the state machine.
    EscrowRecord:
      type: object
      required:
        - escrow_id
        - state
        - corridor
        - amount_usdc
        - created_at
        - state_changed_at
      properties:
        escrow_id:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/EscrowState"
        corridor:
          type: string
          example: "USD_NGN"
        amount_usdc:
          type: string
          description: Decimal USDC amount with 7 decimal places.
          example: "100.0000000"
        created_at:
          type: string
          format: date-time
        state_changed_at:
          type: string
          format: date-time
    EscrowCreateRequest:
      type: object
      required: [sender_account, corridor, amount_usdc]
      properties:
        sender_account:
          type: string
          description: Stellar G… address of the sender.
          example: "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
        corridor:
          type: string
          description: Currency corridor code (e.g. USD_NGN, EUR_GHS, GBP_KES).
          example: "USD_NGN"
        amount_usdc:
          type: number
          format: double
          description: Positive USDC amount to lock in escrow.
          example: 100.0
    EscrowCreateResponse:
      type: object
      required: [escrow_id, state]
      properties:
        escrow_id:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/EscrowState"
    EscrowStateResponse:
      type: object
      required: [escrow_id, state]
      properties:
        escrow_id:
          type: string
          format: uuid
        state:
          $ref: "#/components/schemas/EscrowState"
...
  responses:
    400BadRequest:
      description: Bad request — missing or invalid input.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "sender_account is required"
    401Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Unauthorized"
...
    404NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "escrow not found"
    409Conflict:
      description: Conflict with current resource state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "escrow is in state Released, not Funded"
    500InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Internal server error"
...
paths:
...
  /api/v1/escrow:
    post:
      operationId: createEscrow
      summary: Create an escrow (initiate payment)
      description: >
        Locks USDC in escrow for the specified corridor. Returns an escrow ID
        and initial `Funded` state. The escrow transitions to `Released`,
        `Refundable`, or `Cancelled` via subsequent lifecycle endpoints.
      tags: [Escrow]
      security: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscrowCreateRequest"
            example:
              sender_account: "GAAZI4TCR3TY5OJHCTJC2A4QSY6CJWJH5IAJTGKIN2ER7LBNVKOCCWN"
              corridor: "USD_NGN"
              amount_usdc: 100.0
      responses:
        "201":
          description: Escrow created in `Funded` state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowCreateResponse"
              example:
                escrow_id: "a3bb189e-8bf9-3888-9912-ace4e6543002"
                state: "Funded"
        "400":
          $ref: "#/components/responses/400BadRequest"
        "500":
          $ref: "#/components/responses/500InternalError"
  /api/v1/escrow/{id}:
    get:
      operationId: getEscrow
      summary: Get escrow state
      description: Returns full escrow details including current state and timestamps.
      tags: [Escrow]
      security: []
      parameters:
        - $ref: "#/components/parameters/EscrowId"
      responses:
        "200":
          description: Escrow record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowRecord"
              example:
                escrow_id: "a3bb189e-8bf9-3888-9912-ace4e6543002"
                state: "Funded"
                corridor: "USD_NGN"
                amount_usdc: "100.0000000"
                created_at: "2026-07-26T01:00:00.000Z"
                state_changed_at: "2026-07-26T01:00:00.000Z"
        "404":
          $ref: "#/components/responses/404NotFound"
  /api/v1/escrow/{id}/stream:
    get:
      operationId: streamEscrowEvents
      summary: Real-time escrow timeline via Server-Sent Events
      description: >
        Opens a persistent SSE stream that emits `escrow_event` events for
        every state change. The browser's `EventSource` (or the `Last-Event-ID`
        header / `?lastEventId=` query parameter) is used to resume from the
        last received event after a reconnect — no event is skipped or
        double-delivered.

        A `: heartbeat` comment is sent every 15 seconds to keep idle proxies
        from closing the connection.
      tags: [Escrow]
      security: []
      parameters:
        - $ref: "#/components/parameters/EscrowId"
        - name: Last-Event-ID
          in: header
          required: false
          description: >
            Monotonic event ID from the last successfully received SSE frame.
            Events after this ID are replayed on reconnect.
          schema:
            type: integer
            example: 5
        - name: lastEventId
          in: query
          required: false
          description: Fallback for non-browser SSE clients that cannot send the header.
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream opened.
          headers:
            Content-Type:
              schema:
                type: string
                enum: ["text/event-stream"]
            Cache-Control:
              schema:
                type: string
                enum: ["no-cache, no-transform"]
            X-Accel-Buffering:
              schema:
                type: string
                enum: ["no"]
          content:
            text/event-stream:
              schema:
                type: string
                description: >
                  Each event frame has the form:
                  ```
                  id: <monotonic-int>
                  event: escrow_event
                  data: <EscrowEvent JSON>
                  ```
              example: |
                id: 1
                event: escrow_event
                data: {"escrow_id":"a3bb189e-...","type":"created","state":"Funded","corridor":"USD_NGN","amount_usdc":"100.0000000","occurred_at":"2026-07-26T01:00:00.000Z"}
        "404":
          $ref: "#/components/responses/404NotFound"
  /api/v1/escrow/{id}/release:
    post:
      operationId: releaseEscrow
      summary: Release escrow funds to agent (SEP-10 Ed25519)
      description: >
        Transitions the escrow from `Funded` to `Released`, releasing USDC to
        the off-ramp agent. Requires full SEP-10 Ed25519 JWT (not the
        shared-secret variant).
      tags: [Escrow]
      security:
        - sep10JWT: []
      parameters:
        - $ref: "#/components/parameters/EscrowId"
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Escrow released.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowStateResponse"
              example:
                escrow_id: "a3bb189e-8bf9-3888-9912-ace4e6543002"
                state: "Released"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          $ref: "#/components/responses/409Conflict"
  /api/v1/escrow/{id}/dispute:
    post:
      operationId: disputeEscrow
      summary: Open a dispute on an escrow (SEP-10 Ed25519)
      description: >
        Transitions the escrow from `Funded` or `Released` to `Refundable`,
        marking it for manual review. Requires full SEP-10 Ed25519 JWT.
      tags: [Escrow]
      security:
        - sep10JWT: []
      parameters:
        - $ref: "#/components/parameters/EscrowId"
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dispute opened; escrow is now Refundable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EscrowStateResponse"
              example:
                escrow_id: "a3bb189e-8bf9-3888-9912-ace4e6543002"
                state: "Refundable"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "409":
          $ref: "#/components/responses/409Conflict"

  # ---------------------------------------------------------------------------
  # Content Tiers
  # ---------------------------------------------------------------------------
...